requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
phonenumbers==8.13.4
pycodestyle==2.10.0
pytz==2022.7
sqlparse==0.4.3
